refactor message generation

+++ rosidl_typesupport_opensplice_cpp/rosidl_typesupport_opensplice_cpp/__init__.py
@@ -17,6 +17,7 @@ from io import StringIO
 import os
 import subprocess
 
+from rosidl_cmake import convert_camel_case_to_lower_case_underscore
 from rosidl_cmake import extract_message_types
 from rosidl_parser import parse_message_file
 from rosidl_parser import parse_service_file
@@ -24,12 +25,8 @@ from rosidl_parser import validate_field_types
 
 
 def generate_dds_opensplice_cpp(
-        pkg_name, dds_interface_files, dds_interface_base_path, deps, output_dir, idl_pp):
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
-
+    pkg_name, dds_interface_files, dds_interface_base_path, deps, output_basepath, idl_pp
+):
     include_dirs = [dds_interface_base_path]
     for dep in deps:
         # only take the first : for separation, as Windows follows with a C:\
@@ -42,6 +39,18 @@ def generate_dds_opensplice_cpp(
             include_dirs.append(idl_base_path)
 
     for idl_file in dds_interface_files:
+        # get two level of parent folders for idl file
+        folder = os.path.dirname(idl_file)
+        parent_folder = os.path.dirname(folder)
+        output_path = os.path.join(
+            output_basepath,
+            os.path.basename(parent_folder),
+            os.path.basename(folder))
+        try:
+            os.makedirs(output_path)
+        except FileExistsError:
+            pass
+
         cmd = [idl_pp]
         for include_dir in include_dirs:
             cmd += ['-I', include_dir]
@@ -49,7 +58,7 @@ def generate_dds_opensplice_cpp(
             '-S',
             '-l', 'cpp',
             '-o', 'dds-types',
-            '-d', output_dir,
+            '-d', output_path,
             idl_file
         ]
         subprocess.check_call(cmd)
@@ -61,38 +70,42 @@ def generate_typesupport_opensplice_cpp(
     pkg_name, ros_interface_files, deps, output_dir, template_dir
 ):
     mapping_msgs = {
-        os.path.join(template_dir, 'msg_TypeSupport.h.template'): '%s_TypeSupport.h',
-        os.path.join(template_dir, 'msg_TypeSupport.cpp.template'): '%s_TypeSupport.cpp',
+        os.path.join(template_dir, 'msg__type_support.hpp.template'): '%s__type_support.hpp',
+        os.path.join(template_dir, 'msg__type_support.cpp.template'): '%s__type_support.cpp',
     }
 
     mapping_srvs = {
-        os.path.join(template_dir, 'srv_ServiceTypeSupport.cpp.template'):
-        '%s_ServiceTypeSupport.cpp',
+        os.path.join(template_dir, 'srv__type_support.cpp.template'):
+        '%s__type_support.cpp',
     }
 
     for template_file in mapping_msgs.keys():
-        assert(os.path.exists(template_file))
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
     for template_file in mapping_srvs.keys():
-        assert(os.path.exists(template_file))
-
-    try:
-        os.makedirs(output_dir)
-    except FileExistsError:
-        pass
+        assert os.path.exists(template_file), 'Could not find template: ' + template_file
 
     known_msg_types = extract_message_types(pkg_name, ros_interface_files, deps)
 
+    functions = {
+        'get_header_filename_from_msg_name': convert_camel_case_to_lower_case_underscore,
+    }
+
     for idl_file in ros_interface_files:
-        filename, extension = os.path.splitext(idl_file)
+        extension = os.path.splitext(idl_file)[1]
         if extension == '.msg':
             spec = parse_message_file(pkg_name, idl_file)
             validate_field_types(spec, known_msg_types)
+            subfolder = os.path.basename(os.path.dirname(idl_file))
             for template_file, generated_filename in mapping_msgs.items():
-                generated_file = os.path.join(output_dir, generated_filename % spec.base_type.type)
+                generated_file = os.path.join(
+                    output_dir, subfolder, 'dds_opensplice', generated_filename %
+                    convert_camel_case_to_lower_case_underscore(spec.base_type.type))
 
                 try:
                     output = StringIO()
+                    data = {'spec': spec, 'subfolder': subfolder}
+                    data.update(functions)
                     # TODO reuse interpreter
                     interpreter = em.Interpreter(
                         output=output,
@@ -100,13 +113,14 @@ def generate_typesupport_opensplice_cpp(
                             em.RAW_OPT: True,
                             em.BUFFERED_OPT: True,
                         },
-                        globals={'spec': spec},
+                        globals=data,
                     )
                     interpreter.file(open(template_file))
                     content = output.getvalue()
                     interpreter.shutdown()
                 except Exception:
-                    os.remove(generated_file)
+                    if os.path.exists(generated_file):
+                        os.remove(generated_file)
                     raise
 
                 # only overwrite file if necessary
@@ -114,6 +128,10 @@ def generate_typesupport_opensplice_cpp(
                     with open(generated_file, 'r') as h:
                         if h.read() == content:
                             continue
+                try:
+                    os.makedirs(os.path.dirname(generated_file))
+                except FileExistsError:
+                    pass
                 with open(generated_file, 'w') as h:
                     h.write(content)
 
@@ -121,10 +139,14 @@ def generate_typesupport_opensplice_cpp(
             spec = parse_service_file(pkg_name, idl_file)
             validate_field_types(spec, known_msg_types)
             for template_file, generated_filename in mapping_srvs.items():
-                generated_file = os.path.join(output_dir, generated_filename % spec.srv_name)
+                generated_file = os.path.join(
+                    output_dir, 'srv', 'dds_opensplice', generated_filename %
+                    convert_camel_case_to_lower_case_underscore(spec.srv_name))
 
                 try:
                     output = StringIO()
+                    data = {'spec': spec}
+                    data.update(functions)
                     # TODO reuse interpreter
                     interpreter = em.Interpreter(
                         output=output,
@@ -132,13 +154,14 @@ def generate_typesupport_opensplice_cpp(
                             em.RAW_OPT: True,
                             em.BUFFERED_OPT: True,
                         },
-                        globals={'spec': spec},
+                        globals=data,
                     )
                     interpreter.file(open(template_file))
                     content = output.getvalue()
                     interpreter.shutdown()
                 except Exception:
-                    os.remove(generated_file)
+                    if os.path.exists(generated_file):
+                        os.remove(generated_file)
                     raise
 
                 # only overwrite file if necessary
@@ -146,6 +169,10 @@ def generate_typesupport_opensplice_cpp(
                     with open(generated_file, 'r') as h:
                         if h.read() == content:
                             continue
+                try:
+                    os.makedirs(os.path.dirname(generated_file))
+                except FileExistsError:
+                    pass
                 with open(generated_file, 'w') as h:
                     h.write(content)
